requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cheroot==8.5.2
CherryPy==18.6.0
configparser==5.0.1
Django==3.1.7
django-minio-backend==3.3.2
jaraco.classes==3.2.1
jaraco.collections==3.2.0
jaraco.functools==3.2.1
jaraco.text==3.5.0
more-itertools==8.7.0
portend==2.7.1
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
tempora==4.0.1
urllib3==1.26.3
zc.lockfile==2.0
djangorestframework==3.13.1
drf-yasg==1.21.0
transliterate==1.10.2